functional testing fix and additional test
delete fix:
write all test in one test block to keep the session of the current
user.

additional test:
add assignment under subject. assignment cam be mark as done and
mark as not done

functional testing fix and additional test + code coverage

delete fix:
write all test in one test block to keep the session of the current
user.

additional test:
add assignment under subject. assignment cam be mark as done and
mark as not done

code coverage (incomplete and below 50%):
• register
• logout
• login
• addsubject

+++ laravel_ToDo/app/Http/Controllers/SubjectsController.php
@@ -95,12 +95,12 @@ class SubjectsController extends Controller
         $subject = Subject::find($id);
 
         // double check user
-        // if(auth()->user()->id !== $subject->user_id) {
-        //     return redirect()->back()->with('error!');
-        // }
+        if(auth()->user()->id !== $subject->user_id) {
+            return redirect()->back()->with('error!');
+        }
 
         // delete all assignments under subject
-        // Assignment::where('courseName_id', $subject->id)->delete();
+        Assignment::where('courseName_id', $subject->id)->delete();
 
         $subject->delete();
         return redirect()->back();
